+++ app/controller/admin/editmovie.js
@@ -194,7 +194,7 @@ module.exports = {
                             }
 
                             //altes Cover löschen
-                            if(newMovie.coverImg.length > 0) {
+                            if(newMovie.coverImg.length > 0 && fs.existsSync(__dirname + "/../../public/image/cover/" + newMovie.coverImg)) {
 
                                 fs.unlinkSync(__dirname + "/../../public/image/cover/" + newMovie.coverImg);
                             }
